requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
websockets
sqlmodel
# mysqlclient
pymysql
pydantic-settings
gunicorn
cryptography
requests